Auto-generates Kotlin codecs from Swift `@WireFormat` types via the
`emit-kotlin-codecs` CLI, replacing all hand-written mirror codecs
across SheetMusicAndroid, SheetMusicAudioAndroid, and the Examples
Compose demo.

Phase 1/2/3 + follow-ups A (DrawProgram migration) and B (architectural
cleanup) all landed on this branch — 29 commits.

Key deliverables:
- `Sources/WireFormatKotlinEmitter/` — emits codec objects for
  @WireFormat (struct), @WireFormatChoice (sum-type), @WireFormatEnum
  (raw enum). Supports nested-type refs, cross-package imports,
  `--include-package` CLI filter for per-module codegen.
- `Sources/EmitKotlinCodecs/` — CLI driver with idempotent file writes
  and stale-file sweep.
- `:SheetMusicAndroid` hosts the single canonical
  `io.github.jiyimeta.sheetmusic.wireformat.{BinaryReader,BinaryWriter}`
  consumed transitively by every Kotlin module emitting codecs.
- Each Gradle module's :emitKotlinCodecs task declares the package(s)
  it owns; clean disjoint partition, no dex-merge excludes.
- Audio side: ~600 LOC of hand-written decoders deleted; only
  GMInstrumentDecoder remains (external wire layout).
- Examples app: 122-line DrawProgramDecoder.kt replaced by tiny
  DrawProgramReader that delegates structural decode + validates
  magic/version.

Verified: swift test 1404/250 suites · :SheetMusicAndroid /
:SheetMusicAudioAndroid assembleRelease + testDebugUnitTest ·
Examples/Android :app assembleDebug + 8 unit tests · installed on
Pixel_6_Pro_API_36 (emulator-5556) and Pixel_6_API_34 (emulator-5554).

Spec:  docs/superpowers/specs/2026-05-23-kotlin-codec-codegen-design.md
Plan:  docs/superpowers/plans/2026-05-23-kotlin-codec-codegen.md

+++ Android/SheetMusicAudioAndroid/src/main/kotlin/io/github/jiyimeta/sheetmusic/audio/serialization/GMInstrumentDecoder.kt
@@ -1,6 +1,7 @@
 package io.github.jiyimeta.sheetmusic.audio.serialization
 
 import io.github.jiyimeta.sheetmusic.audio.model.GMInstrument
+import io.github.jiyimeta.sheetmusic.wireformat.BinaryReader
 
 /**
  * Decodes the `[GMInstrument]` blob produced by Swift's
@@ -28,7 +29,7 @@ internal object GMInstrumentDecoder {
         val count = r.readI32()
         val out = ArrayList<GMInstrument>(count)
         for (i in 0 until count) {
-            val program = r.readU8()
+            val program = r.readU8().toInt()
             val nameLen = r.readI32()
             require(nameLen >= 0) { "GMInstrument name length is negative: $nameLen" }
             val nameBytes = ByteArray(nameLen)
@@ -36,7 +37,7 @@ internal object GMInstrumentDecoder {
                 nameBytes[j] = r.readU8().toByte()
             }
             val name = String(nameBytes, Charsets.UTF_8)
-            val familyOrdinal = r.readU8()
+            val familyOrdinal = r.readU8().toInt()
             out.add(
                 GMInstrument(
                     program = program,
